+++ cpp/dolfinx/refinement/refine.h
@@ -6,8 +6,11 @@
 
 #pragma once
 
-#include <cstdint>
-#include <span>
+#include "plaza.h"
+#include "refine.h"
+#include <dolfinx/common/IndexMap.h>
+#include <dolfinx/common/log.h>
+#include <dolfinx/mesh/Mesh.h>
 
 namespace dolfinx::mesh
 {
@@ -23,8 +26,28 @@ namespace dolfinx::refinement
 /// @param[in] redistribute If `true` refined mesh is re-partitioned
 /// across MPI ranks.
 /// @return Refined mesh
-mesh::Mesh<double> refine(const mesh::Mesh<double>& mesh,
-                          bool redistribute = true);
+template <typename T>
+mesh::Mesh<T> refine(const mesh::Mesh<T>& mesh, bool redistribute = true)
+{
+  if (mesh.topology().cell_type() != mesh::CellType::triangle
+      and mesh.topology().cell_type() != mesh::CellType::tetrahedron)
+  {
+    throw std::runtime_error("Refinement only defined for simplices");
+  }
+
+  auto [refined_mesh, parent_cell, parent_facet]
+      = plaza::refine(mesh, redistribute, plaza::Option::none);
+
+  // Report the number of refined cells
+  const int D = mesh.topology().dim();
+  const std::int64_t n0 = mesh.topology().index_map(D)->size_global();
+  const std::int64_t n1 = refined_mesh.topology().index_map(D)->size_global();
+  LOG(INFO) << "Number of cells increased from " << n0 << " to " << n1 << " ("
+            << 100.0 * (static_cast<double>(n1) / static_cast<double>(n0) - 1.0)
+            << "%% increase).";
+
+  return refined_mesh;
+}
 
 /// @brief Create a locally refined mesh.
 ///
@@ -36,7 +59,29 @@ mesh::Mesh<double> refine(const mesh::Mesh<double>& mesh,
 /// @param[in] redistribute If `true` refined mesh is re-partitioned
 /// across MPI ranks.
 /// @return Refined mesh.
-mesh::Mesh<double> refine(const mesh::Mesh<double>& mesh,
-                          std::span<const std::int32_t> edges,
-                          bool redistribute = true);
+template <typename T>
+mesh::Mesh<T> refine(const mesh::Mesh<T>& mesh,
+                     std::span<const std::int32_t> edges,
+                     bool redistribute = true)
+{
+  if (mesh.topology().cell_type() != mesh::CellType::triangle
+      and mesh.topology().cell_type() != mesh::CellType::tetrahedron)
+  {
+    throw std::runtime_error("Refinement only defined for simplices");
+  }
+
+  auto [refined_mesh, parent_cell, parent_facet]
+      = plaza::refine(mesh, edges, redistribute, plaza::Option::none);
+
+  // Report the number of refined cells
+  const int D = mesh.topology().dim();
+  const std::int64_t n0 = mesh.topology().index_map(D)->size_global();
+  const std::int64_t n1 = refined_mesh.topology().index_map(D)->size_global();
+  LOG(INFO) << "Number of cells increased from " << n0 << " to " << n1 << " ("
+            << 100.0 * (static_cast<double>(n1) / static_cast<double>(n0) - 1.0)
+            << "%% increase).";
+
+  return refined_mesh;
+}
+
 } // namespace dolfinx::refinement
